% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlee.gen.R
\name{clean_data_hash}
\alias{clean_data_hash}
\title{Make plotly data identifier reproducible}
\usage{
clean_data_hash(p)
}
\arguments{
\item{p}{\link[plotly:plot_ly]{Plotly object} to modify.}
}
\value{
A \link[plotly:plot_ly]{plotly object}.
}
\description{
Replaces the random \link[plotly:plotly_data]{data} identifier of a \link[plotly:plot_ly]{plotly object} with a deterministic hash value of the data's content.
}
\details{
This function is especially useful to apply before a plotly object or its JSON representation (as generated for rendered R Markdown / Quarto documents)
is \href{https://en.wikipedia.org/wiki/Version_control}{version controlled} (e.g. via Git).

To also make the htmlwidget identifier reproducible when rendering a Plotly chart, use \code{\link[htmlwidgets:setWidgetIdSeed]{htmlwidgets::setWidgetIdSeed()}}.
}
\examples{
p <- plotly::plot_ly(data = mtcars,
                     x = ~mpg,
                     type = "histogram")

p2 <- plotly::plot_ly(data = mtcars,
                      x = ~mpg,
                      type = "histogram")

p_deterministic <- plotlee::clean_data_hash(p)
p2_deterministic <- plotlee::clean_data_hash(p)

p$x$cur_data
p2$x$cur_data
p_deterministic$x$cur_data
p2_deterministic$x$cur_data
}
