% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlee.gen.R
\name{simplify_trace_ids}
\alias{simplify_trace_ids}
\title{Make plotly trace identifiers reproducible}
\usage{
simplify_trace_ids(p)
}
\arguments{
\item{p}{\link[plotly:plot_ly]{Plotly object} to modify.}
}
\value{
A \link[plotly:plot_ly]{plotly object}.
}
\description{
Replaces the random trace identifiers of a \link[plotly:plot_ly]{plotly object} with simple sequential ones (\code{trace_0}, \code{trace_1} and so on).
}
\details{
This function is especially useful to apply before a plotly object or its JSON representation (as generated for rendered R Markdown / Quarto documents)
is \href{https://en.wikipedia.org/wiki/Version_control}{version controlled} (e.g. via Git).

To also make the htmlwidget identifier reproducible when rendering a Plotly chart, use \code{\link[htmlwidgets:setWidgetIdSeed]{htmlwidgets::setWidgetIdSeed()}}.
}
\examples{
p <- plotly::plot_ly(data = mtcars,
                     x = ~mpg,
                     type = "histogram")

p2 <- plotly::plot_ly(data = mtcars,
                      x = ~mpg,
                      type = "histogram")

p_deterministic <- plotlee::simplify_trace_ids(p)
p2_deterministic <- plotlee::simplify_trace_ids(p2)

names(p$x$visdat)
names(p2$x$visdat)
names(p_deterministic$x$visdat)
names(p2_deterministic$x$visdat)
}
