% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlee-package.R
\docType{package}
\name{plotlee-package}
\alias{plotlee}
\alias{plotlee-package}
\title{plotlee: Unofficial Helper Functions around Plotly}
\description{
Provides unofficial helper functions around the plotly package, the R API for the JavaScript graphing library plotly.js.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/rpkg.dev/plotlee}
  \item Report bugs at \url{https://gitlab.com/rpkg.dev/plotlee/issues}
}

}
\author{
\strong{Maintainer}: Salim Brüggemann \email{salim-b@pm.me} (\href{https://orcid.org/0000-0002-5329-5987}{ORCID})

}
\keyword{internal}
